rename package to match pattern name
