+++ design-patterns/creational-patterns/src/factoryMethod/Archiver.java
@@ -1,4 +1,4 @@
-package factory;
+package factoryMethod;
 
 public abstract class Archiver {
