+++ design-patterns/creational-patterns/src/factoryMethod/RarFileArchive.java
@@ -1,4 +1,4 @@
-package factory;
+package factoryMethod;
 
 import java.io.File;
