+++ design-patterns/creational-patterns/src/factoryMethod/WindowsArchiver.java
@@ -1,4 +1,4 @@
-package factory;
+package factoryMethod;
 
 public class WindowsArchiver extends Archiver {
